Code Style Check Review
1- Unecessary rules

https://codenarc.github.io/CodeNarc/codenarc-rules-unnecessary.html#unnecessaryobjectreferences-rule
For an object initialization, instead of populating their properties in
multiple rules, use the tap definition that receives a closure with the
actions expected to run inside the instance. In the end tap returns the
object.

2- Use of instanceof

https://codenarc.github.io/CodeNarc/codenarc-rules-design.html#instanceof-rule
Using instanceof can be considered a code smell, as you would have to update all your conditionals whenever you add a new class.
But I think that is not the case:
result[k] = result[k] instanceof Map ? merge(result[k], v) : v
The visitor pattern is so much expensive than we pretend in this case.

3- Static string replay

https://codenarc.github.io/CodeNarc/codenarc-rules-dry.html#duplicatestringliteral-rule
'repos' is repeated during the code.
I changed it to _repos as a constant field.
I used the same notation in DockerCompose.

4- String value code style

For a string initialization with a static value (that don't depend on
other variables) use ' ' instead of " ".

Signed-off-by: Samuel Bernardo <samuel@lip.pt>